Minor updates
Removed a few unnecessary items from the CMakeLists.txt files to facilitate compilation. Modified the implementation of small matrices to allow them to be mere wrapper of pre-allocated memory blocks, which is useful for their use as slices out of a 3rd order tensor, which is now available as small_tensor.